fix review comment - 2
