+++ src/Server/RestRouterHandlers/RestRouterHandler.h
@@ -36,38 +36,7 @@ public:
     /// a correct `http_status` code will be set by trying best.
     /// This function may throw, and caller will need catch the exception
     /// and sends back HTTP `500` to clients
-    void execute(HTTPServerRequest & request, HTTPServerResponse & response)
-    {
-        setupHTTPContext(request);
-
-        Int32 http_status = HTTPResponse::HTTP_OK;
-        String response_payload;
-
-        if (streamingInput())
-        {
-            response_payload = execute(request.getStream(), http_status);
-        }
-        else
-        {
-            auto payload{parseRequest(request)};
-            if (streamingOutput())
-            {
-                /// For keep-alive to work.
-                if (request.getVersion() == HTTPServerRequest::HTTP_1_1)
-                    response.setChunkedTransferEncoding(true);
-                execute(payload, response);
-            }
-            else
-                response_payload = execute(payload, http_status);
-        }
-
-        /// Only send response_payload when streamingOutput is false, otherwise it will cause output messed up when compressison is enabled.
-        if (!streamingOutput())
-        {
-            response.setStatusAndReason(HTTPResponse::HTTPStatus(http_status));
-            *response.send() << response_payload << std::endl;
-        }
-    }
+    void execute(HTTPServerRequest & request, HTTPServerResponse & response);
 
     const String & getPathParameter(const String & name, const String & default_value = "") const
     {
@@ -80,21 +49,6 @@ public:
         {
             return default_value;
         }
-    }
-
-    static Poco::JSON::Object::Ptr parseRequest(HTTPServerRequest & request)
-    {
-        String data = "{}";
-
-        auto size = request.getContentLength();
-        if (size > 0)
-        {
-            data.resize(size);
-            request.getStream().readStrict(data.data(), size);
-        }
-
-        Poco::JSON::Parser parser;
-        return parser.parse(data).extract<Poco::JSON::Object::Ptr>();
     }
 
     void setPathParameter(const String & name, const String & value) { path_parameters[name] = value; }
