+++ src/Server/ya.make
@@ -33,12 +33,12 @@ SRCS(
     ProtocolServerAdapter.cpp
     ReplicasStatusHandler.cpp
     RestHTTPRequestHandler.cpp
-    RestRouterHandlers/ColumnRestRouterHandler.cpp
     RestRouterHandlers/DatabaseRestRouterHandler.cpp
     RestRouterHandlers/IngestRawStoreHandler.cpp
     RestRouterHandlers/IngestRestRouterHandler.cpp
     RestRouterHandlers/IngestStatusHandler.cpp
     RestRouterHandlers/RawstoreTableRestRouterHandler.cpp
+    RestRouterHandlers/RestRouterHandler.cpp
     RestRouterHandlers/RestStatusHandler.cpp
     RestRouterHandlers/SQLAnalyzerRestRouterHandler.cpp
     RestRouterHandlers/SchemaValidator.cpp
